move checkpoint restoration and checking supported spaces into test_utils. Make supported spaces test also support tf2 out of the box
